requirements.txt
python-escpos[all]
python-dotenv
